+++ pkg/traces/config_test.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/grafana/agent/pkg/traces/pushreceiver"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/config"
@@ -89,6 +90,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -103,7 +105,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -130,6 +132,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -156,7 +159,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["attributes", "batch"]
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -174,6 +177,7 @@ remote_write:
       password_file: ` + passwordFile.Name(),
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -192,7 +196,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -211,6 +215,7 @@ remote_write:
       password_file: ` + passwordFileExtraNewline.Name(),
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -229,7 +234,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -244,6 +249,7 @@ remote_write:
     endpoint: example.com:12345`,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -260,7 +266,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -276,6 +282,7 @@ remote_write:
     compression: none`,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -292,7 +299,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -313,6 +320,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -333,7 +341,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -380,6 +388,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -396,7 +405,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -425,6 +434,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -454,7 +464,7 @@ service:
     traces:
       exporters: ["otlp/1", "otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -472,6 +482,7 @@ batch:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -490,7 +501,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["batch"]
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -512,6 +523,7 @@ spanmetrics:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   noop:
   jaeger:
     protocols:
@@ -538,7 +550,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["spanmetrics"]
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
     metrics/spanmetrics:
       exporters: ["remote_write"]
       receivers: ["noop"]
@@ -558,6 +570,7 @@ spanmetrics:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   noop:
   jaeger:
     protocols:
@@ -579,7 +592,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["spanmetrics"]
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
     metrics/spanmetrics:
       exporters: ["prometheus"]
       receivers: ["noop"]
@@ -634,6 +647,7 @@ tail_sampling:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -685,7 +699,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["tail_sampling"]
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -719,6 +733,7 @@ receivers:
   jaeger:
     protocols:
       grpc:
+  push_receiver: {}
   otlp/lb:
     protocols:
       grpc:
@@ -760,7 +775,7 @@ service:
     traces/0:
       exporters: ["loadbalancing"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["jaeger", "push_receiver"]
     traces/1:
       exporters: ["otlp/0"]
       processors: ["tail_sampling"]
@@ -781,6 +796,7 @@ automatic_logging:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -799,7 +815,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["automatic_logging"]
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
       `,
 		},
 		{
@@ -819,6 +835,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -838,7 +855,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -856,6 +873,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -875,7 +893,7 @@ service:
     traces:
       exporters: ["otlphttp/0", "otlp/1"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -894,6 +912,7 @@ prom_sd_operation_type: update
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -913,7 +932,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["prom_sd_processor"]
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -930,6 +949,7 @@ service_graphs:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -946,7 +966,7 @@ service:
     traces:
       exporters: ["otlp/0"]
       processors: ["service_graphs"]
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -963,6 +983,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -979,7 +1000,7 @@ service:
     traces:
       exporters: ["jaeger/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -1000,6 +1021,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -1018,7 +1040,7 @@ service:
     traces:
       exporters: ["jaeger/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -1038,6 +1060,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -1061,7 +1084,7 @@ service:
     traces:
       exporters: ["jaeger/0", "otlp/1"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -1101,6 +1124,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -1125,7 +1149,7 @@ service:
     traces:
       exporters: ["otlphttp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -1152,6 +1176,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -1181,7 +1206,7 @@ service:
     traces:
       exporters: ["otlphttp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -1211,6 +1236,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+ push_receiver: {}
  jaeger:
    protocols:
      grpc:
@@ -1248,7 +1274,7 @@ service:
     traces:
       exporters: ["otlphttp/0", "otlp/1"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 		{
@@ -1273,6 +1299,7 @@ remote_write:
 `,
 			expectedConfig: `
 receivers:
+  push_receiver: {}
   jaeger:
     protocols:
       grpc:
@@ -1301,7 +1328,7 @@ service:
     traces:
       exporters: ["otlphttp/0"]
       processors: []
-      receivers: ["jaeger"]
+      receivers: ["push_receiver", "jaeger"]
 `,
 		},
 	}
@@ -1681,6 +1708,20 @@ receivers:
 	assert.True(t, strings.Contains(string(data), "<secret>"))
 }
 
+func TestCreatingPushReceiver(t *testing.T) {
+	test := `
+receivers:
+  jaeger:
+    protocols:
+      grpc:`
+	cfg := InstanceConfig{}
+	err := yaml.Unmarshal([]byte(test), &cfg)
+	assert.Nil(t, err)
+	otel, err := cfg.otelConfig()
+	assert.Nil(t, err)
+	assert.Contains(t, otel.Service.Pipelines[config.NewComponentID("traces")].Receivers, config.NewComponentID(pushreceiver.TypeStr))
+}
+
 // sortPipelines is a helper function to lexicographically sort a pipeline's exporters
 func sortPipelines(cfg *config.Config) {
 	tracePipeline := cfg.Pipelines[config.NewComponentID(config.TracesDataType)]
@@ -1688,11 +1729,13 @@ func sortPipelines(cfg *config.Config) {
 		return
 	}
 	var (
-		exp  = tracePipeline.Exporters
-		recv = tracePipeline.Receivers
-		ext  = cfg.Service.Extensions
+		exp         = tracePipeline.Exporters
+		recv        = tracePipeline.Receivers
+		ext         = cfg.Service.Extensions
+		serviceRecv = cfg.Service.Pipelines[config.NewComponentID(config.TracesDataType)].Receivers
 	)
 	sort.Slice(exp, func(i, j int) bool { return exp[i].String() > exp[j].String() })
 	sort.Slice(recv, func(i, j int) bool { return recv[i].String() > recv[j].String() })
 	sort.Slice(ext, func(i, j int) bool { return ext[i].String() > ext[j].String() })
+	sort.Slice(serviceRecv, func(i, j int) bool { return serviceRecv[i].String() > serviceRecv[j].String() })
 }